#Issue 968 (#1008)
* Move test_grid to correct folder, initial updates on tests

* TST: Updated tests for test_grid

* TST: Move file test_md_grid to new location

* TST: Updated tests in test_md_grid

* TST: Small change in documentation

* STY: black, flake8

* Move test_grid to correct folder, initial updates on tests

* TST: Updated tests for test_grid

* TST: Move file test_md_grid to new location

* TST: Updated tests in test_md_grid

* TST: Small change in documentation

* STY: black, flake8

* STY: black, flake8 and isort

---------
